Fix isdefined() to output function call literally with variables expanded
isdefined() is not an evaluating function - it should be output literally
with variable substitution. Removed from function registry so it's treated
as an unknown function that gets output as-is.

This allows isdefined(@var) to output 'isdefined(value)' instead of
evaluating to true/false.

Note: Full fixture 063-type-functions-defined test requires ruleset
variable support which is not yet implemented.